feat: add complexity tests
