+++ router-tests/complexity_limits_test.go
@@ -1,0 +1,581 @@
+package integration
+
+import (
+	"github.com/stretchr/testify/require"
+	"github.com/wundergraph/cosmo/router-tests/testenv"
+	"github.com/wundergraph/cosmo/router/pkg/config"
+	"github.com/wundergraph/cosmo/router/pkg/otel"
+	"github.com/wundergraph/cosmo/router/pkg/trace/tracetest"
+	"go.opentelemetry.io/otel/sdk/metric"
+	"go.opentelemetry.io/otel/sdk/trace"
+	tracetest2 "go.opentelemetry.io/otel/sdk/trace/tracetest"
+	"net/http"
+	"testing"
+)
+
+func TestComplexityLimits(t *testing.T) {
+	t.Parallel()
+	t.Run("old max query depth configuration still works", func(t *testing.T) {
+		t.Run("max query depth of 0 doesn't block", func(t *testing.T) {
+			t.Parallel()
+			testenv.Run(t, &testenv.Config{
+				ModifySecurityConfiguration: func(securityConfiguration *config.SecurityConfiguration) {
+					if securityConfiguration.DepthLimit == nil {
+						securityConfiguration.DepthLimit = &config.QueryDepthConfiguration{}
+					}
+					securityConfiguration.DepthLimit.Enabled = true
+					securityConfiguration.DepthLimit.Limit = 0
+					securityConfiguration.DepthLimit.CacheSize = 1024
+				},
+			}, func(t *testing.T, xEnv *testenv.Environment) {
+				res := xEnv.MakeGraphQLRequestOK(testenv.GraphQLRequest{
+					Query: `{ employee(id:1) { id details { forename surname } } }`,
+				})
+				require.JSONEq(t, `{"data":{"employee":{"id":1,"details":{"forename":"Jens","surname":"Neuse"}}}}`, res.Body)
+			})
+		})
+
+		t.Run("allows queries up to the max depth", func(t *testing.T) {
+			t.Parallel()
+			testenv.Run(t, &testenv.Config{
+				ModifySecurityConfiguration: func(securityConfiguration *config.SecurityConfiguration) {
+					if securityConfiguration.DepthLimit == nil {
+						securityConfiguration.DepthLimit = &config.QueryDepthConfiguration{}
+					}
+					securityConfiguration.DepthLimit.Enabled = true
+					securityConfiguration.DepthLimit.Limit = 3
+					securityConfiguration.DepthLimit.CacheSize = 1024
+				},
+			}, func(t *testing.T, xEnv *testenv.Environment) {
+				res := xEnv.MakeGraphQLRequestOK(testenv.GraphQLRequest{
+					Query: `{ employee(id:1) { id details { forename surname } } }`,
+				})
+				require.JSONEq(t, `{"data":{"employee":{"id":1,"details":{"forename":"Jens","surname":"Neuse"}}}}`, res.Body)
+			})
+		})
+
+		t.Run("max query depth blocks queries over the limit", func(t *testing.T) {
+			t.Parallel()
+			testenv.Run(t, &testenv.Config{
+				ModifySecurityConfiguration: func(securityConfiguration *config.SecurityConfiguration) {
+					if securityConfiguration.DepthLimit == nil {
+						securityConfiguration.DepthLimit = &config.QueryDepthConfiguration{}
+					}
+					securityConfiguration.DepthLimit.Enabled = true
+					securityConfiguration.DepthLimit.Limit = 2
+					securityConfiguration.DepthLimit.CacheSize = 1024
+				},
+			}, func(t *testing.T, xEnv *testenv.Environment) {
+				res, _ := xEnv.MakeGraphQLRequest(testenv.GraphQLRequest{
+					Query: `{ employee(id:1) { id details { forename surname } } }`,
+				})
+				require.Equal(t, 400, res.Response.StatusCode)
+				require.Equal(t, `{"errors":[{"message":"The query depth 3 exceeds the max query depth allowed (2)"}]}`, res.Body)
+			})
+		})
+
+		t.Run("max query depth blocks persisted queries over the limit", func(t *testing.T) {
+			t.Parallel()
+			testenv.Run(t, &testenv.Config{
+				ModifySecurityConfiguration: func(securityConfiguration *config.SecurityConfiguration) {
+					if securityConfiguration.DepthLimit == nil {
+						securityConfiguration.DepthLimit = &config.QueryDepthConfiguration{}
+					}
+					securityConfiguration.DepthLimit.Enabled = true
+					securityConfiguration.DepthLimit.Limit = 2
+					securityConfiguration.DepthLimit.CacheSize = 1024
+				},
+			}, func(t *testing.T, xEnv *testenv.Environment) {
+				header := make(http.Header)
+				header.Add("graphql-client-name", "my-client")
+				res, _ := xEnv.MakeGraphQLRequestOverGET(testenv.GraphQLRequest{
+					OperationName: []byte(`Find`),
+					Variables:     []byte(`{"criteria":  {"nationality":  "GERMAN"   }}`),
+					Extensions:    []byte(`{"persistedQuery": {"version": 1, "sha256Hash": "e33580cf6276de9a75fb3b1c4b7580fec2a1c8facd13f3487bf6c7c3f854f7e3"}}`),
+					Header:        header,
+				})
+				require.Equal(t, 400, res.Response.StatusCode)
+				require.Equal(t, `{"errors":[{"message":"The query depth 3 exceeds the max query depth allowed (2)"}]}`, res.Body)
+			})
+		})
+
+		t.Run("max query depth doesn't block persisted queries if DisableDepthLimitPersistedOperations set", func(t *testing.T) {
+			t.Parallel()
+			testenv.Run(t, &testenv.Config{
+				ModifySecurityConfiguration: func(securityConfiguration *config.SecurityConfiguration) {
+					if securityConfiguration.DepthLimit == nil {
+						securityConfiguration.DepthLimit = &config.QueryDepthConfiguration{}
+					}
+					securityConfiguration.DepthLimit.Enabled = true
+					securityConfiguration.DepthLimit.Limit = 2
+					securityConfiguration.DepthLimit.CacheSize = 1024
+					securityConfiguration.DepthLimit.IgnorePersistedOperations = true
+				},
+			}, func(t *testing.T, xEnv *testenv.Environment) {
+				header := make(http.Header)
+				header.Add("graphql-client-name", "my-client")
+				res, _ := xEnv.MakeGraphQLRequestOverGET(testenv.GraphQLRequest{
+					OperationName: []byte(`Find`),
+					Variables:     []byte(`{"criteria":  {"nationality":  "GERMAN"   }}`),
+					Extensions:    []byte(`{"persistedQuery": {"version": 1, "sha256Hash": "e33580cf6276de9a75fb3b1c4b7580fec2a1c8facd13f3487bf6c7c3f854f7e3"}}`),
+					Header:        header,
+				})
+				require.Equal(t, 200, res.Response.StatusCode)
+				require.Equal(t, `{"data":{"findEmployees":[{"id":1,"details":{"forename":"Jens","surname":"Neuse"}},{"id":2,"details":{"forename":"Dustin","surname":"Deus"}},{"id":4,"details":{"forename":"Björn","surname":"Schwenzer"}},{"id":11,"details":{"forename":"Alexandra","surname":"Neuse"}}]}}`, res.Body)
+			})
+		})
+
+		t.Run("query depth validation caches success and failure runs", func(t *testing.T) {
+			t.Parallel()
+
+			metricReader := metric.NewManualReader()
+			exporter := tracetest.NewInMemoryExporter(t)
+			testenv.Run(t, &testenv.Config{
+				TraceExporter: exporter,
+				MetricReader:  metricReader,
+				ModifySecurityConfiguration: func(securityConfiguration *config.SecurityConfiguration) {
+					if securityConfiguration.DepthLimit == nil {
+						securityConfiguration.DepthLimit = &config.QueryDepthConfiguration{}
+					}
+					securityConfiguration.DepthLimit.Enabled = true
+					securityConfiguration.DepthLimit.Limit = 2
+					securityConfiguration.DepthLimit.CacheSize = 1024
+				},
+			}, func(t *testing.T, xEnv *testenv.Environment) {
+				failedRes, _ := xEnv.MakeGraphQLRequest(testenv.GraphQLRequest{
+					Query: `{ employee(id:1) { id details { forename surname } } }`,
+				})
+				require.Equal(t, 400, failedRes.Response.StatusCode)
+				require.Equal(t, `{"errors":[{"message":"The query depth 3 exceeds the max query depth allowed (2)"}]}`, failedRes.Body)
+
+				testSpan := requireSpanWithName(t, exporter, "Operation - Validate")
+				require.Contains(t, testSpan.Attributes(), otel.WgQueryDepth.Int(3))
+				require.Contains(t, testSpan.Attributes(), otel.WgQueryDepthCacheHit.Bool(false))
+				exporter.Reset()
+
+				failedRes2, _ := xEnv.MakeGraphQLRequest(testenv.GraphQLRequest{
+					Query: `{ employee(id:1) { id details { forename surname } } }`,
+				})
+				require.Equal(t, 400, failedRes2.Response.StatusCode)
+				require.Equal(t, `{"errors":[{"message":"The query depth 3 exceeds the max query depth allowed (2)"}]}`, failedRes2.Body)
+
+				testSpan2 := requireSpanWithName(t, exporter, "Operation - Validate")
+				require.Contains(t, testSpan2.Attributes(), otel.WgQueryDepth.Int(3))
+				require.Contains(t, testSpan2.Attributes(), otel.WgQueryDepthCacheHit.Bool(true))
+				exporter.Reset()
+
+				successRes := xEnv.MakeGraphQLRequestOK(testenv.GraphQLRequest{
+					Query: `query { employees { id } }`,
+				})
+				require.JSONEq(t, employeesIDData, successRes.Body)
+				testSpan3 := requireSpanWithName(t, exporter, "Operation - Validate")
+				require.Contains(t, testSpan3.Attributes(), otel.WgQueryDepth.Int(2))
+				require.Contains(t, testSpan3.Attributes(), otel.WgQueryDepthCacheHit.Bool(false))
+				exporter.Reset()
+
+				successRes2 := xEnv.MakeGraphQLRequestOK(testenv.GraphQLRequest{
+					Query: `query { employees { id } }`,
+				})
+				require.JSONEq(t, employeesIDData, successRes2.Body)
+				testSpan4 := requireSpanWithName(t, exporter, "Operation - Validate")
+				require.Contains(t, testSpan4.Attributes(), otel.WgQueryDepth.Int(2))
+				require.Contains(t, testSpan4.Attributes(), otel.WgQueryDepthCacheHit.Bool(true))
+			})
+		})
+	})
+
+	t.Run("depth limit", func(t *testing.T) {
+		t.Run("depth limit blocks queries over the limit", func(t *testing.T) {
+			t.Parallel()
+			testenv.Run(t, &testenv.Config{
+				ModifySecurityConfiguration: func(securityConfiguration *config.SecurityConfiguration) {
+					securityConfiguration.ComplexityLimits = &config.ComplexityLimits{
+						Depth: &config.ComplexityLimit{
+							Enabled: true,
+							Limit:   2,
+						},
+					}
+				},
+			}, func(t *testing.T, xEnv *testenv.Environment) {
+				res, _ := xEnv.MakeGraphQLRequest(testenv.GraphQLRequest{
+					Query: `{ employee(id:1) { id details { forename surname } } }`,
+				})
+				require.Equal(t, 400, res.Response.StatusCode)
+				require.Equal(t, `{"errors":[{"message":"The query depth 3 exceeds the max query depth allowed (2)"}]}`, res.Body)
+			})
+		})
+
+		t.Run("depth limit blocks persisted queries over the limit", func(t *testing.T) {
+			t.Parallel()
+			testenv.Run(t, &testenv.Config{
+				ModifySecurityConfiguration: func(securityConfiguration *config.SecurityConfiguration) {
+					securityConfiguration.ComplexityLimits = &config.ComplexityLimits{
+						Depth: &config.ComplexityLimit{
+							Enabled: true,
+							Limit:   2,
+						},
+					}
+				},
+			}, func(t *testing.T, xEnv *testenv.Environment) {
+				header := make(http.Header)
+				header.Add("graphql-client-name", "my-client")
+				res, _ := xEnv.MakeGraphQLRequestOverGET(testenv.GraphQLRequest{
+					OperationName: []byte(`Find`),
+					Variables:     []byte(`{"criteria":  {"nationality":  "GERMAN"   }}`),
+					Extensions:    []byte(`{"persistedQuery": {"version": 1, "sha256Hash": "e33580cf6276de9a75fb3b1c4b7580fec2a1c8facd13f3487bf6c7c3f854f7e3"}}`),
+					Header:        header,
+				})
+				require.Equal(t, 400, res.Response.StatusCode)
+				require.Equal(t, `{"errors":[{"message":"The query depth 3 exceeds the max query depth allowed (2)"}]}`, res.Body)
+			})
+		})
+
+		t.Run("depth limit doesn't block persisted queries if IgnorePersistedOperations set", func(t *testing.T) {
+			t.Parallel()
+			testenv.Run(t, &testenv.Config{
+				ModifySecurityConfiguration: func(securityConfiguration *config.SecurityConfiguration) {
+					securityConfiguration.ComplexityLimits = &config.ComplexityLimits{
+						Depth: &config.ComplexityLimit{
+							Enabled:                   true,
+							Limit:                     2,
+							IgnorePersistedOperations: true,
+						},
+					}
+				},
+			}, func(t *testing.T, xEnv *testenv.Environment) {
+				header := make(http.Header)
+				header.Add("graphql-client-name", "my-client")
+				res, _ := xEnv.MakeGraphQLRequestOverGET(testenv.GraphQLRequest{
+					OperationName: []byte(`Find`),
+					Variables:     []byte(`{"criteria":  {"nationality":  "GERMAN"   }}`),
+					Extensions:    []byte(`{"persistedQuery": {"version": 1, "sha256Hash": "e33580cf6276de9a75fb3b1c4b7580fec2a1c8facd13f3487bf6c7c3f854f7e3"}}`),
+					Header:        header,
+				})
+				require.Equal(t, 200, res.Response.StatusCode)
+				require.Equal(t, `{"data":{"findEmployees":[{"id":1,"details":{"forename":"Jens","surname":"Neuse"}},{"id":2,"details":{"forename":"Dustin","surname":"Deus"}},{"id":4,"details":{"forename":"Björn","surname":"Schwenzer"}},{"id":11,"details":{"forename":"Alexandra","surname":"Neuse"}}]}}`, res.Body)
+			})
+		})
+	})
+
+	t.Run("total fields limit", func(t *testing.T) {
+		t.Parallel()
+
+		t.Run("total fields limit blocks queries over the limit", func(t *testing.T) {
+			testenv.Run(t, &testenv.Config{
+				ModifySecurityConfiguration: func(securityConfiguration *config.SecurityConfiguration) {
+					securityConfiguration.ComplexityLimits = &config.ComplexityLimits{
+						TotalFields: &config.ComplexityLimit{
+							Enabled: true,
+							Limit:   1,
+						},
+					}
+				},
+			}, func(t *testing.T, xEnv *testenv.Environment) {
+				res, _ := xEnv.MakeGraphQLRequest(testenv.GraphQLRequest{
+					Query: `{ employee(id:1) { id details { forename surname } } }`,
+				})
+				require.Equal(t, 400, res.Response.StatusCode)
+				require.Equal(t, `{"errors":[{"message":"The query total fields 2 exceeds the max total fields allowed (1)"}]}`, res.Body)
+			})
+		})
+
+		t.Run("total fields allows queries under the limit", func(t *testing.T) {
+			testenv.Run(t, &testenv.Config{
+				ModifySecurityConfiguration: func(securityConfiguration *config.SecurityConfiguration) {
+					securityConfiguration.ComplexityLimits = &config.ComplexityLimits{
+						TotalFields: &config.ComplexityLimit{
+							Enabled: true,
+							Limit:   3,
+						},
+					}
+				},
+			}, func(t *testing.T, xEnv *testenv.Environment) {
+				res := xEnv.MakeGraphQLRequestOK(testenv.GraphQLRequest{
+					Query: `{ employee(id:1) { id details { forename surname } } }`,
+				})
+				require.JSONEq(t, `{"data":{"employee":{"id":1,"details":{"forename":"Jens","surname":"Neuse"}}}}`, res.Body)
+			})
+		})
+
+		t.Run("total fields caches success and failure runs", func(t *testing.T) {
+			t.Parallel()
+
+			metricReader := metric.NewManualReader()
+			exporter := tracetest.NewInMemoryExporter(t)
+			testenv.Run(t, &testenv.Config{
+				TraceExporter: exporter,
+				MetricReader:  metricReader,
+				ModifySecurityConfiguration: func(securityConfiguration *config.SecurityConfiguration) {
+					securityConfiguration.ComplexityCalculationCache = &config.ComplexityCalculationCache{
+						Enabled:   true,
+						CacheSize: 1024,
+					}
+					securityConfiguration.ComplexityLimits = &config.ComplexityLimits{
+						TotalFields: &config.ComplexityLimit{
+							Enabled: true,
+							Limit:   1,
+						},
+					}
+				},
+			}, func(t *testing.T, xEnv *testenv.Environment) {
+				failedRes, _ := xEnv.MakeGraphQLRequest(testenv.GraphQLRequest{
+					Query: `{ employee(id:1) { id details { forename surname } } }`,
+				})
+				require.Equal(t, 400, failedRes.Response.StatusCode)
+				require.Equal(t, `{"errors":[{"message":"The query total fields 2 exceeds the max total fields allowed (1)"}]}`, failedRes.Body)
+
+				testSpan := requireSpanWithName(t, exporter, "Operation - Validate")
+				require.Contains(t, testSpan.Attributes(), otel.WgQueryTotalFields.Int(2))
+				require.Contains(t, testSpan.Attributes(), otel.WgQueryDepthCacheHit.Bool(false))
+				exporter.Reset()
+
+				failedRes2, _ := xEnv.MakeGraphQLRequest(testenv.GraphQLRequest{
+					Query: `{ employee(id:1) { id details { forename surname } } }`,
+				})
+				require.Equal(t, 400, failedRes2.Response.StatusCode)
+				require.Equal(t, `{"errors":[{"message":"The query total fields 2 exceeds the max total fields allowed (1)"}]}`, failedRes2.Body)
+
+				testSpan2 := requireSpanWithName(t, exporter, "Operation - Validate")
+				require.Contains(t, testSpan2.Attributes(), otel.WgQueryTotalFields.Int(2))
+				require.Contains(t, testSpan2.Attributes(), otel.WgQueryDepthCacheHit.Bool(true))
+				exporter.Reset()
+
+				successRes := xEnv.MakeGraphQLRequestOK(testenv.GraphQLRequest{
+					Query: `query { employees { id } }`,
+				})
+				require.JSONEq(t, employeesIDData, successRes.Body)
+				testSpan3 := requireSpanWithName(t, exporter, "Operation - Validate")
+				require.Contains(t, testSpan3.Attributes(), otel.WgQueryTotalFields.Int(1))
+				require.Contains(t, testSpan3.Attributes(), otel.WgQueryDepthCacheHit.Bool(false))
+				exporter.Reset()
+
+				successRes2 := xEnv.MakeGraphQLRequestOK(testenv.GraphQLRequest{
+					Query: `query { employees { id } }`,
+				})
+				require.JSONEq(t, employeesIDData, successRes2.Body)
+				testSpan4 := requireSpanWithName(t, exporter, "Operation - Validate")
+				require.Contains(t, testSpan4.Attributes(), otel.WgQueryTotalFields.Int(1))
+				require.Contains(t, testSpan4.Attributes(), otel.WgQueryDepthCacheHit.Bool(true))
+			})
+		})
+	})
+
+	t.Run("root fields limit", func(t *testing.T) {
+		t.Parallel()
+
+		t.Run("root fields limit blocks queries over the limit", func(t *testing.T) {
+			testenv.Run(t, &testenv.Config{
+				ModifySecurityConfiguration: func(securityConfiguration *config.SecurityConfiguration) {
+					securityConfiguration.ComplexityLimits = &config.ComplexityLimits{
+						RootFields: &config.ComplexityLimit{
+							Enabled: true,
+							Limit:   2,
+						},
+					}
+				},
+			}, func(t *testing.T, xEnv *testenv.Environment) {
+				res, _ := xEnv.MakeGraphQLRequest(testenv.GraphQLRequest{
+					Query: `query { initialPayload employee(id:1) { id } employees { id } }`,
+				})
+				require.Equal(t, 400, res.Response.StatusCode)
+				require.Equal(t, `{"errors":[{"message":"The number of root fields 3 exceeds the root field limit allowed (2)"}]}`, res.Body)
+			})
+		})
+
+		t.Run("root fields allows queries under the limit", func(t *testing.T) {
+			testenv.Run(t, &testenv.Config{
+				ModifySecurityConfiguration: func(securityConfiguration *config.SecurityConfiguration) {
+					securityConfiguration.ComplexityLimits = &config.ComplexityLimits{
+						RootFields: &config.ComplexityLimit{
+							Enabled: true,
+							Limit:   2,
+						},
+					}
+				},
+			}, func(t *testing.T, xEnv *testenv.Environment) {
+				res := xEnv.MakeGraphQLRequestOK(testenv.GraphQLRequest{
+					Query: `query { employee(id:1) { id } }`,
+				})
+				require.JSONEq(t, `{"data":{"employee":{"id":1}}}`, res.Body)
+			})
+		})
+
+		t.Run("root fields caches success and failure runs", func(t *testing.T) {
+			t.Parallel()
+
+			metricReader := metric.NewManualReader()
+			exporter := tracetest.NewInMemoryExporter(t)
+			testenv.Run(t, &testenv.Config{
+				TraceExporter: exporter,
+				MetricReader:  metricReader,
+				ModifySecurityConfiguration: func(securityConfiguration *config.SecurityConfiguration) {
+					securityConfiguration.ComplexityCalculationCache = &config.ComplexityCalculationCache{
+						Enabled:   true,
+						CacheSize: 1024,
+					}
+					securityConfiguration.ComplexityLimits = &config.ComplexityLimits{
+						RootFields: &config.ComplexityLimit{
+							Enabled: true,
+							Limit:   2,
+						},
+					}
+				},
+			}, func(t *testing.T, xEnv *testenv.Environment) {
+				failedRes, _ := xEnv.MakeGraphQLRequest(testenv.GraphQLRequest{
+					Query: `query { initialPayload employee(id:1) { id } employees { id } }`,
+				})
+				require.Equal(t, 400, failedRes.Response.StatusCode)
+				require.Equal(t, `{"errors":[{"message":"The number of root fields 3 exceeds the root field limit allowed (2)"}]}`, failedRes.Body)
+
+				testSpan := requireSpanWithName(t, exporter, "Operation - Validate")
+				require.Contains(t, testSpan.Attributes(), otel.WgQueryRootFields.Int(3))
+				require.Contains(t, testSpan.Attributes(), otel.WgQueryDepthCacheHit.Bool(false))
+				exporter.Reset()
+
+				failedRes2, _ := xEnv.MakeGraphQLRequest(testenv.GraphQLRequest{
+					Query: `query { initialPayload employee(id:1) { id } employees { id } }`,
+				})
+				require.Equal(t, 400, failedRes2.Response.StatusCode)
+				require.Equal(t, `{"errors":[{"message":"The number of root fields 3 exceeds the root field limit allowed (2)"}]}`, failedRes2.Body)
+
+				testSpan2 := requireSpanWithName(t, exporter, "Operation - Validate")
+				require.Contains(t, testSpan2.Attributes(), otel.WgQueryRootFields.Int(3))
+				require.Contains(t, testSpan2.Attributes(), otel.WgQueryDepthCacheHit.Bool(true))
+				exporter.Reset()
+
+				successRes := xEnv.MakeGraphQLRequestOK(testenv.GraphQLRequest{
+					Query: `query { employees { id } }`,
+				})
+				require.JSONEq(t, employeesIDData, successRes.Body)
+				testSpan3 := requireSpanWithName(t, exporter, "Operation - Validate")
+				require.Contains(t, testSpan3.Attributes(), otel.WgQueryRootFields.Int(1))
+				require.Contains(t, testSpan3.Attributes(), otel.WgQueryDepthCacheHit.Bool(false))
+				exporter.Reset()
+
+				successRes2 := xEnv.MakeGraphQLRequestOK(testenv.GraphQLRequest{
+					Query: `query { employees { id } }`,
+				})
+				require.JSONEq(t, employeesIDData, successRes2.Body)
+				testSpan4 := requireSpanWithName(t, exporter, "Operation - Validate")
+				require.Contains(t, testSpan4.Attributes(), otel.WgQueryRootFields.Int(1))
+				require.Contains(t, testSpan4.Attributes(), otel.WgQueryDepthCacheHit.Bool(true))
+			})
+		})
+	})
+
+	t.Run("root field aliases limit", func(t *testing.T) {
+		t.Parallel()
+
+		t.Run("root field aliases limit blocks queries over the limit", func(t *testing.T) {
+			testenv.Run(t, &testenv.Config{
+				ModifySecurityConfiguration: func(securityConfiguration *config.SecurityConfiguration) {
+					securityConfiguration.ComplexityLimits = &config.ComplexityLimits{
+						RootFieldAliases: &config.ComplexityLimit{
+							Enabled: true,
+							Limit:   1,
+						},
+					}
+				},
+			}, func(t *testing.T, xEnv *testenv.Environment) {
+				res, _ := xEnv.MakeGraphQLRequest(testenv.GraphQLRequest{
+					Query: `query { firstemployee: employee(id:1) { id } employee2: employee(id:2) { id } }`,
+				})
+				require.Equal(t, 400, res.Response.StatusCode)
+				require.Equal(t, `{"errors":[{"message":"The number of root fields aliases 2 exceeds the root field aliases limit allowed (1)"}]}`, res.Body)
+			})
+		})
+
+		t.Run("root field aliases allows queries under the limit", func(t *testing.T) {
+			testenv.Run(t, &testenv.Config{
+				ModifySecurityConfiguration: func(securityConfiguration *config.SecurityConfiguration) {
+					securityConfiguration.ComplexityLimits = &config.ComplexityLimits{
+						RootFieldAliases: &config.ComplexityLimit{
+							Enabled: true,
+							Limit:   2,
+						},
+					}
+				},
+			}, func(t *testing.T, xEnv *testenv.Environment) {
+				res := xEnv.MakeGraphQLRequestOK(testenv.GraphQLRequest{
+					Query: `query { firstemployee: employee(id:1) { id } employee2: employee(id:2) { id } }`,
+				})
+				require.Equal(t, `{"data":{"firstemployee":{"id":1},"employee2":{"id":2}}}`, res.Body)
+			})
+		})
+
+		t.Run("root fields caches success and failure runs", func(t *testing.T) {
+			t.Parallel()
+
+			metricReader := metric.NewManualReader()
+			exporter := tracetest.NewInMemoryExporter(t)
+			testenv.Run(t, &testenv.Config{
+				TraceExporter: exporter,
+				MetricReader:  metricReader,
+				ModifySecurityConfiguration: func(securityConfiguration *config.SecurityConfiguration) {
+					securityConfiguration.ComplexityCalculationCache = &config.ComplexityCalculationCache{
+						Enabled:   true,
+						CacheSize: 1024,
+					}
+					securityConfiguration.ComplexityLimits = &config.ComplexityLimits{
+						RootFieldAliases: &config.ComplexityLimit{
+							Enabled: true,
+							Limit:   1,
+						},
+					}
+				},
+			}, func(t *testing.T, xEnv *testenv.Environment) {
+				failedRes, _ := xEnv.MakeGraphQLRequest(testenv.GraphQLRequest{
+					Query: `query { firstemployee: employee(id:1) { id } employee2: employee(id:2) { id } }`,
+				})
+				require.Equal(t, 400, failedRes.Response.StatusCode)
+				require.Equal(t, `{"errors":[{"message":"The number of root fields aliases 2 exceeds the root field aliases limit allowed (1)"}]}`, failedRes.Body)
+
+				testSpan := requireSpanWithName(t, exporter, "Operation - Validate")
+				require.Contains(t, testSpan.Attributes(), otel.WgQueryRootFieldAliases.Int(2))
+				require.Contains(t, testSpan.Attributes(), otel.WgQueryDepthCacheHit.Bool(false))
+				exporter.Reset()
+
+				failedRes2, _ := xEnv.MakeGraphQLRequest(testenv.GraphQLRequest{
+					Query: `query { firstemployee: employee(id:1) { id } employee2: employee(id:2) { id } }`,
+				})
+				require.Equal(t, 400, failedRes2.Response.StatusCode)
+				require.Equal(t, `{"errors":[{"message":"The number of root fields aliases 2 exceeds the root field aliases limit allowed (1)"}]}`, failedRes2.Body)
+
+				testSpan2 := requireSpanWithName(t, exporter, "Operation - Validate")
+				require.Contains(t, testSpan2.Attributes(), otel.WgQueryRootFieldAliases.Int(2))
+				require.Contains(t, testSpan2.Attributes(), otel.WgQueryDepthCacheHit.Bool(true))
+				exporter.Reset()
+
+				successRes := xEnv.MakeGraphQLRequestOK(testenv.GraphQLRequest{
+					Query: `query { employees { id } }`,
+				})
+				require.JSONEq(t, employeesIDData, successRes.Body)
+				testSpan3 := requireSpanWithName(t, exporter, "Operation - Validate")
+				require.Contains(t, testSpan3.Attributes(), otel.WgQueryRootFieldAliases.Int(0))
+				require.Contains(t, testSpan3.Attributes(), otel.WgQueryDepthCacheHit.Bool(false))
+				exporter.Reset()
+
+				successRes2 := xEnv.MakeGraphQLRequestOK(testenv.GraphQLRequest{
+					Query: `query { employees { id } }`,
+				})
+				require.JSONEq(t, employeesIDData, successRes2.Body)
+				testSpan4 := requireSpanWithName(t, exporter, "Operation - Validate")
+				require.Contains(t, testSpan4.Attributes(), otel.WgQueryRootFieldAliases.Int(0))
+				require.Contains(t, testSpan4.Attributes(), otel.WgQueryDepthCacheHit.Bool(true))
+			})
+		})
+	})
+}
+
+func requireSpanWithName(t *testing.T, exporter *tracetest2.InMemoryExporter, name string) trace.ReadOnlySpan {
+	sn := exporter.GetSpans().Snapshots()
+	var testSpan trace.ReadOnlySpan
+	for _, span := range sn {
+		if span.Name() == name {
+			testSpan = span
+			break
+		}
+	}
+	require.NotNil(t, testSpan)
+	return testSpan
+}
