--- router-tests/testenv/complexity_limits_test.go
@@ -1,1 +1,0 @@
-package testenv
